+++ .travis.yml
@@ -1,3 +1,5 @@
+dist: trusty
+
 language: node_js
 
 node_js:
