+++ src/app/blender/blender.component.jsx
@@ -2,10 +2,14 @@ import React from 'react';
 import { connect } from 'react-redux';
 import * as keywords from './data/keywords.json';
 import * as domains from './data/domains.json';
+import * as objectifs from './data/objectives.json';
 import CheckboxKeyword from './checkboxKeyword/checkboxKeyword.component';
 import CheckboxDomain from './checkboxDomain/checkboxDomain.component';
+import CheckboxObjectif from './checkboxObjectif/checkboxObjectif.component';
 import Mixeur from './mixeur/mixeur.component';
 import { addKeyword } from './blender.action';
+
+import './blender.css';
 
 const Form = () => {
   const showKeyword = keywords.map(item => (
@@ -14,13 +18,28 @@ const Form = () => {
   const showTheme = domains.map(item => (
     <CheckboxDomain item={item} key={item.id} />
   ));
+  const showObjectif = objectifs.map(item => (
+    <CheckboxObjectif item={item} key={item.id} />
+  ));
   return (
     <div className="form">
-      <h2>Thème des conférences</h2>
-      {showTheme}
-      <h2>Mots clefs</h2>
-      {showKeyword}
-      <Mixeur />
+      <div className="items">
+        <div className="item">
+          <h1 className="category-title">Les thématiques</h1>
+          {showTheme}
+        </div>
+        <div className="item">
+          <h1 className="category-title">Les sujets</h1>
+          {showKeyword}
+        </div>
+        <div className="item">
+          <h1 className="category-title">Vos objectifs</h1>
+          {showObjectif}
+        </div>
+        <div className="item">
+          <Mixeur />
+        </div>
+      </div>
     </div>
   );
 };
